Revert "Chunked upload"
This reverts commit 6d76de75c2c7c7c3b412abcaac11c30a9d3c8412.
