+++ app/components/admin/upload/AudioFileUpload.tsx
@@ -4,11 +4,15 @@ import type { FC } from 'react';
 import { useCallback, useEffect, useRef, useState } from 'react';
 import { useControlField, useField } from 'remix-validated-form';
 
+import { UPLOAD_AUDIO_FORM_KEY } from '~/forms/upload-audio';
 import { useSse } from '~/hooks/useSse';
 import type { loader as audioUploadLoader } from '~/routes/admin.audio-upload.$id';
+import type { action as newAudioUploadAction } from '~/routes/admin.audio-upload.new';
 import type { AudioFileProcessingEvent } from '~/sse.server/audio-file-events';
 
-import type { useUploadAudioFile } from './useUploadAudioFile';
+import { xhrPromise } from './xhrPromise';
+
+type UploadResponse = SerializeFrom<typeof newAudioUploadAction>;
 
 type SerializeFrom<T> = ReturnType<typeof useLoaderData<T>>;
 
@@ -29,12 +33,14 @@ function displayConversionStatus(
 
 interface AudioFileUploadProps {
   name: string;
-  uploadAudioFile: ReturnType<typeof useUploadAudioFile>;
+  isUploading: boolean;
+  setIsUploading: (isUploading: boolean) => void;
 }
 
 export const AudioFileUpload: FC<AudioFileUploadProps> = ({
   name,
-  uploadAudioFile,
+  isUploading,
+  setIsUploading,
 }) => {
   const { getInputProps } = useField(name);
   const [fileId, setFileId] = useControlField<string | undefined>(name);
@@ -66,31 +72,41 @@ export const AudioFileUpload: FC<AudioFileUploadProps> = ({
   const file = fileState ?? fetcher.data;
 
   const [fileName, setFileName] = useState<string>();
+  const [uploadProgress, setUploadProgress] = useState(0);
+
   const fileInputRef = useRef<HTMLInputElement>(null);
-
-  const { start, pause, resume, abort, state } = uploadAudioFile;
-
-  useEffect(() => {
-    if (state?.status === 'done') {
-      setFileState(state.file);
-
-      // Wait a bit to make sure the fetcher is not triggered before the
-      // file upload state is updated.
-      setTimeout(() => {
-        setFileId(state.file.id);
-      }, 100);
-    }
-  }, [setFileId, state]);
-
   const onUploadClick = () => {
     const fileInput = fileInputRef.current;
     if (!fileInput?.files?.length) return;
 
     const file = fileInput.files[0]!;
     fileInput.value = '';
+    setIsUploading(true);
     setFileName(file.name);
+    setUploadProgress(0);
 
-    start(file);
+    const form = new FormData();
+    form.append(UPLOAD_AUDIO_FORM_KEY, file);
+    xhrPromise(form, {
+      url: '/admin/audio-upload/new',
+      onProgress: setUploadProgress,
+      errorOnBadStatus: true,
+    })
+      .then((response) => {
+        const file = JSON.parse(response.responseText) as UploadResponse;
+
+        setFileState(file);
+
+        // Wait a bit to make sure the fetcher is not triggered before the
+        // file upload state is updated.
+        setTimeout(() => {
+          setFileId(file.id);
+          setIsUploading(false);
+        }, 100);
+      })
+      .catch((error: unknown) => {
+        console.error(`Audio file upload ${name} failed.`, error);
+      });
   };
 
   const onRemoveFileClick = () => {
@@ -107,55 +123,39 @@ export const AudioFileUpload: FC<AudioFileUploadProps> = ({
         })}
       />
       <p>
-        {state === undefined ? (
+        {isUploading ? (
+          <>
+            Uploading… <progress value={uploadProgress} /> {fileName}
+          </>
+        ) : fileId ? (
+          file ? (
+            file.conversionStatus === 'DONE' ? (
+              <>
+                Duration: {file.duration}{' '}
+                <button type="button" onClick={onRemoveFileClick}>
+                  Remove file
+                </button>
+              </>
+            ) : file.errorMessage ? (
+              <>Error while converting audio file: {file.errorMessage}</>
+            ) : (
+              <>
+                {displayConversionStatus(file.conversionStatus)}{' '}
+                {file.conversionProgress === null ? null : (
+                  <progress value={file.conversionProgress} />
+                )}
+              </>
+            )
+          ) : (
+            'Loading…'
+          )
+        ) : (
           <>
             <input type="file" ref={fileInputRef} accept="audio/*" />{' '}
             <button type="button" onClick={onUploadClick}>
               Upload
             </button>
           </>
-        ) : state.status === 'in progress' ? (
-          <>
-            Uploading… <progress value={state.progress} /> {fileName}{' '}
-            <button type="button" onClick={pause}>
-              Pause
-            </button>{' '}
-            <button type="button" onClick={abort}>
-              Cancel
-            </button>
-          </>
-        ) : state.status === 'paused' ? (
-          <>
-            Paused <progress value={state.progress} /> {fileName}{' '}
-            <button type="button" onClick={resume}>
-              Resume
-            </button>{' '}
-            <button type="button" onClick={abort}>
-              Cancel
-            </button>
-          </>
-        ) : state.status === 'error' ? (
-          <>Error while uploading file</>
-        ) : file ? (
-          file.conversionStatus === 'DONE' ? (
-            <>
-              Duration: {file.duration}{' '}
-              <button type="button" onClick={onRemoveFileClick}>
-                Remove file
-              </button>
-            </>
-          ) : file.errorMessage ? (
-            <>Error while converting audio file: {file.errorMessage}</>
-          ) : (
-            <>
-              {displayConversionStatus(file.conversionStatus)}{' '}
-              {file.conversionProgress === null ? null : (
-                <progress value={file.conversionProgress} />
-              )}
-            </>
-          )
-        ) : (
-          'Loading…'
         )}
       </p>
     </>
