--- app/forms/upload.server.ts
@@ -1,1 +1,0 @@
-export const UPLOAD_DIRECTORY = 'upload';
